IDO/IDO/requirements.txt
ipdb==0.13.13
line_profiler==4.2.0
matplotlib==3.10.0
numpy==1.24.1
opencv_python==4.8.0.76
Pillow==10.1.0
Pillow==11.1.0
scipy==1.15.1
timm==0.9.7
torch==2.0.1
torchvision==0.15.2
yacs==0.1.8
